refactor: delete board confirmation dialog
